refactor(agents): Moves agent functions to InstanceAgentsInternal
