+++ src/agents/activate-agent.ts
@@ -6,10 +6,10 @@
  * Licensed under MIT License (see https://github.com/regal/regal)
  */
 
-import { GameInstance } from "../../state";
-import { Agent } from "../agent";
-import { buildActiveAgentArrayProxy, buildActiveAgentProxy } from "../impl";
-import { propertyIsAgentId } from "../instance-agents";
+import { GameInstance } from "../state";
+import { Agent } from "./agent";
+import { buildActiveAgentArrayProxy, buildActiveAgentProxy } from "./impl";
+import { propertyIsAgentId } from "./instance-agents-internal";
 
 /**
  * Returns an activated agent or agent array within the current game context.
