+++ src/agents/index.ts
@@ -5,12 +5,9 @@
  * Licensed under MIT License (see https://github.com/regal/regal)
  */
 
-export { InstanceAgentsInternal } from "./instance-agents";
+export { InstanceAgentsInternal } from "./instance-agents-internal";
 export { Agent, isAgent } from "./agent";
 export { StaticAgentRegistry } from "./static-agent-registry";
 export { PropertyChange, PropertyOperation } from "./agent-properties";
-export { activateAgent } from "./func/activate-agent";
-export { scrubAgents } from "./func/agent-scrub";
-export { recycleInstanceAgents } from "./func/recycle-instance-agents";
+export { activateAgent } from "./activate-agent";
 export { buildActiveAgentProxy, buildInstanceAgents } from "./impl";
-export { buildRevertFunction } from "./func/agent-revert";
